Added a local minimum finder as part of the code; spits out generation where the lowest minimum occurs.

+++ node.java
@@ -12,7 +12,7 @@ import java.math.*;
  */
 public class node {
     public String nodeName;
-    public float weight;
+    public double weight;
     public String calcType;
     public double tier;
     public double error;
